requirements.txt
fastapi~=0.118.2
uvicorn~=0.37.0
httpx~=0.28.1
colorama~=0.4.6
pyyaml~=6.0.2
starlette~=0.48.0
pydantic~=2.12.0
pyjwt~=2.10.1
cryptography~=46.0.2
uvloop~=0.21.0; sys_platform == 'linux'
filelock~=3.20.0
bcrypt~=5.0.0
python-multipart~=0.0.20
contextvars~=2.4
Pillow~=11.3.0
ldap3~=2.9.1
orjson~=3.11.3
python3-saml==1.16.0